+++ DisplayEmployee.java
@@ -2,46 +2,50 @@
 import javax.swing.JFrame;
 import javax.swing.JTextArea;
 import java.awt.Font;
+
 import javax.swing.JPanel;
 import javax.swing.BorderFactory;
+import javax.swing.JButton;
+
 import java.awt.Color;
 import java.util.ArrayList;
+import java.awt.event.*;
 
-public class DisplayEmployee {
+public class DisplayEmployee extends JFrame implements ActionListener {
+    private JButton back = new JButton("Back");
+    private JPanel panel = new JPanel();
 
     // constructor
     public DisplayEmployee() 
     {
-        System.out.println(EmployeeDatabase.getNumberOfEmployees());
+
         // create a GUI frame to display employeeInfo()
-        JFrame frame = new JFrame("Employee Info");
-        frame.setSize(900, 500);
-        frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        frame.setLayout(null);
+        
+        setSize(750, 500);
+        setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
+        setLayout(null);
         
         // create a text area to display employeeInfo()
         JTextArea textArea = new JTextArea();
-        textArea.setBounds(10, 50, 900, 500);
+        textArea.setBounds(10, 50, 900, 300);
         textArea.setEditable(false);
         textArea.setFont(new Font("Times New Roman", Font.PLAIN, 14));
 
-        // System.out.println(getNumberOfEmployees());
+
         //get employee info from EmployeeDatabase
         ArrayList<String> employeeInfo = EmployeeDatabase.getEmployees();
-        System.out.println(employeeInfo);
 
         // display employee info in textArea
         for (int i = 0; i < employeeInfo.size(); i++) 
         {
             textArea.append(employeeInfo.get(i));
         }
-         
 
        
         // // create a panel to display textArea
-        JPanel panel = new JPanel();
+        
         panel.add(textArea);
-        panel.setBounds(0, 0, 900, 500);
+        panel.setBounds(0, 0, 750, 500);
         panel.setLayout(null);
         panel.setFont(new Font("Times New Roman", Font.PLAIN, 30));
         panel.setBorder(BorderFactory.createTitledBorder(BorderFactory.createEtchedBorder(),
@@ -49,11 +53,27 @@ public class DisplayEmployee {
         panel.setBackground(Color.WHITE);
 
         // add the panel to the frame
-        frame.add(panel);
+
+        // add a back button
+        back.setSize(100,30);
+        back.setLocation(100,410);
+        back.setForeground(Color.WHITE);
+        back.setBackground(Color.BLUE);
+        back.addActionListener(this);
+        panel.add(back);
+        add(panel);
+        
         
         // set the frame to exit when closed
-        frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        frame.setVisible(true);
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setVisible(true);
     }
 
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        if(e.getSource() == back ){
+            dispose();
+            new WelcomePage().setVisible(true);
+    }
+}
 }
